+++ maven-plugin/src/main/java/com/marvinformatics/formatter/ConfigurationSource.java
@@ -1,4 +1,4 @@
-package com.relativitas.maven.plugins.formatter;
+package com.marvinformatics.formatter;
 
 import java.io.File;
 import java.nio.charset.Charset;
