+++ maven-plugin/src/main/java/com/marvinformatics/formatter/java/JavaFormatter.java
@@ -1,4 +1,4 @@
-package com.relativitas.maven.plugins.formatter.java;
+package com.marvinformatics.formatter.java;
 
 import java.io.IOException;
 import java.util.Map;
@@ -11,10 +11,10 @@ import org.eclipse.jface.text.Document;
 import org.eclipse.jface.text.IDocument;
 import org.eclipse.text.edits.TextEdit;
 
-import com.relativitas.maven.plugins.formatter.AbstractCacheableFormatter;
-import com.relativitas.maven.plugins.formatter.ConfigurationSource;
-import com.relativitas.maven.plugins.formatter.Formatter;
-import com.relativitas.maven.plugins.formatter.LineEnding;
+import com.marvinformatics.formatter.AbstractCacheableFormatter;
+import com.marvinformatics.formatter.ConfigurationSource;
+import com.marvinformatics.formatter.Formatter;
+import com.marvinformatics.formatter.LineEnding;
 
 public class JavaFormatter extends AbstractCacheableFormatter implements
 		Formatter {
