+++ maven-plugin/src/main/java/com/marvinformatics/formatter/model/Setting.java
@@ -1,4 +1,4 @@
-package com.relativitas.maven.plugins.formatter.xml;
+package com.marvinformatics.formatter.model;
 
 /*
  * Copyright 2010. All work is copyrighted to their respective author(s),
