+++ maven-plugin/src/test/java/com/marvinformatics/formatter/AbstractFormatterTest.java
@@ -1,4 +1,4 @@
-package com.relativitas.maven.plugins.formatter;
+package com.marvinformatics.formatter;
 
 import java.io.File;
 import java.io.IOException;
@@ -15,6 +15,10 @@ import org.apache.maven.plugin.logging.SystemStreamLog;
 
 import com.google.common.base.Charsets;
 import com.google.common.io.Files;
+import com.marvinformatics.formatter.ConfigurationSource;
+import com.marvinformatics.formatter.Formatter;
+import com.marvinformatics.formatter.LineEnding;
+import com.marvinformatics.formatter.Result;
 
 public abstract class AbstractFormatterTest extends TestCase {
