+++ maven-plugin/src/test/java/com/marvinformatics/formatter/LineEndingTest.java
@@ -1,4 +1,4 @@
-package com.relativitas.maven.plugins.formatter;
+package com.marvinformatics.formatter;
 
 /*
  * Copyright 2010. All work is copyrighted to their respective author(s),
